requirements.txt
fastapi
uvicorn
python-multipart
coqui-tts[codec]
edge-tts
torch
torchaudio
transformers>=5.0.0
numpy>=2.0.0
moviepy==1.0.3
proglog
aiofiles